Add files via upload
In case this is helpful for building the synthesizer: PS08 and PS09 in C Programming last year was creating a synthesizer. It's not MIDI-mapped, but is using the computer keyboard. The PDFs for each are quite helpful in understanding what's going on, but let me know if you have any questions.